# supabase_openapi.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "title": "standard public schema",
    "version": "14.5"
  },
  "host": "fwrkrpmqmqxlgvyzrizq.supabase.co:443",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "produces": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "paths": {
...
    "/appointments": {
      "get": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.appointments.id"
          },
          {
            "$ref": "#/parameters/rowFilter.appointments.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.appointments.doctor_id"
          },
          {
            "$ref": "#/parameters/rowFilter.appointments.patient_uid"
          },
          {
            "$ref": "#/parameters/rowFilter.appointments.patient_name"
          },
          {
            "$ref": "#/parameters/rowFilter.appointments.appointment_date"
          },
          {
            "$ref": "#/parameters/rowFilter.appointments.start_time"
          },
          {
            "$ref": "#/parameters/rowFilter.appointments.end_time"
          },
          {
            "$ref": "#/parameters/rowFilter.appointments.status"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/order"
          },
          {
            "$ref": "#/parameters/range"
          },
          {
            "$ref": "#/parameters/rangeUnit"
          },
          {
            "$ref": "#/parameters/offset"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/preferCount"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/appointments"
              },
              "type": "array"
            }
          },
          "206": {
            "description": "Partial Content"
          }
        },
        "tags": [
          "appointments"
        ]
      },
      "post": {
        "parameters": [
          {
            "$ref": "#/parameters/body.appointments"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/preferPost"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "tags": [
          "appointments"
        ]
      },
      "delete": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.appointments.id"
          },
          {
            "$ref": "#/parameters/rowFilter.appointments.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.appointments.doctor_id"
          },
          {
            "$ref": "#/parameters/rowFilter.appointments.patient_uid"
          },
          {
            "$ref": "#/parameters/rowFilter.appointments.patient_name"
          },
          {
            "$ref": "#/parameters/rowFilter.appointments.appointment_date"
          },
          {
            "$ref": "#/parameters/rowFilter.appointments.start_time"
          },
          {
            "$ref": "#/parameters/rowFilter.appointments.end_time"
          },
          {
            "$ref": "#/parameters/rowFilter.appointments.status"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "appointments"
        ]
      },
      "patch": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.appointments.id"
          },
          {
            "$ref": "#/parameters/rowFilter.appointments.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.appointments.doctor_id"
          },
          {
            "$ref": "#/parameters/rowFilter.appointments.patient_uid"
          },
          {
            "$ref": "#/parameters/rowFilter.appointments.patient_name"
          },
          {
            "$ref": "#/parameters/rowFilter.appointments.appointment_date"
          },
          {
            "$ref": "#/parameters/rowFilter.appointments.start_time"
          },
          {
            "$ref": "#/parameters/rowFilter.appointments.end_time"
          },
          {
            "$ref": "#/parameters/rowFilter.appointments.status"
          },
          {
            "$ref": "#/parameters/body.appointments"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "appointments"
        ]
      }
    },
...
    "/attempts": {
      "get": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.attempts.id"
          },
          {
            "$ref": "#/parameters/rowFilter.attempts.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.attempts.patient_uid"
          },
          {
            "$ref": "#/parameters/rowFilter.attempts.item_id"
          },
          {
            "$ref": "#/parameters/rowFilter.attempts.alphabet_name"
          },
          {
            "$ref": "#/parameters/rowFilter.attempts.level_key"
          },
          {
            "$ref": "#/parameters/rowFilter.attempts.score"
          },
          {
            "$ref": "#/parameters/rowFilter.attempts.attempted_at"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/order"
          },
          {
            "$ref": "#/parameters/range"
          },
          {
            "$ref": "#/parameters/rangeUnit"
          },
          {
            "$ref": "#/parameters/offset"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/preferCount"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/attempts"
              },
              "type": "array"
            }
          },
          "206": {
            "description": "Partial Content"
          }
        },
        "tags": [
          "attempts"
        ]
      },
      "post": {
        "parameters": [
          {
            "$ref": "#/parameters/body.attempts"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/preferPost"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "tags": [
          "attempts"
        ]
      },
      "delete": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.attempts.id"
          },
          {
            "$ref": "#/parameters/rowFilter.attempts.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.attempts.patient_uid"
          },
          {
            "$ref": "#/parameters/rowFilter.attempts.item_id"
          },
          {
            "$ref": "#/parameters/rowFilter.attempts.alphabet_name"
          },
          {
            "$ref": "#/parameters/rowFilter.attempts.level_key"
          },
          {
            "$ref": "#/parameters/rowFilter.attempts.score"
          },
          {
            "$ref": "#/parameters/rowFilter.attempts.attempted_at"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "attempts"
        ]
      },
      "patch": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.attempts.id"
          },
          {
            "$ref": "#/parameters/rowFilter.attempts.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.attempts.patient_uid"
          },
          {
            "$ref": "#/parameters/rowFilter.attempts.item_id"
          },
          {
            "$ref": "#/parameters/rowFilter.attempts.alphabet_name"
          },
          {
            "$ref": "#/parameters/rowFilter.attempts.level_key"
          },
          {
            "$ref": "#/parameters/rowFilter.attempts.score"
          },
          {
            "$ref": "#/parameters/rowFilter.attempts.attempted_at"
          },
          {
            "$ref": "#/parameters/body.attempts"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "attempts"
        ]
      }
    },
...
  },
  "definitions": {
...
    "appointments": {
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "description": "Note:\nThis is a Primary Key.<pk/>",
          "format": "bigint",
          "type": "integer"
        },
        "created_at": {
          "default": "now()",
          "format": "timestamp with time zone",
          "type": "string"
        },
        "doctor_id": {
          "description": "Note:\nThis is a Foreign Key to `therapists.id`.<fk table='therapists' column='id'/>",
          "format": "text",
          "type": "string"
        },
        "patient_uid": {
          "format": "text",
          "type": "string"
        },
        "patient_name": {
          "format": "text",
          "type": "string"
        },
        "appointment_date": {
          "format": "date",
          "type": "string"
        },
        "start_time": {
          "format": "time without time zone",
          "type": "string"
        },
        "end_time": {
          "format": "time without time zone",
          "type": "string"
        },
        "status": {
          "default": "confirmed",
          "format": "text",
          "type": "string"
        }
      },
      "type": "object"
    },
...
    "attempts": {
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "description": "Note:\nThis is a Primary Key.<pk/>",
          "format": "bigint",
          "type": "integer"
        },
        "created_at": {
          "default": "now()",
          "format": "timestamp with time zone",
          "type": "string"
        },
        "patient_uid": {
          "format": "text",
          "type": "string"
        },
        "item_id": {
          "format": "text",
          "type": "string"
        },
        "alphabet_name": {
          "format": "text",
          "type": "string"
        },
        "level_key": {
          "format": "text",
          "type": "string"
        },
        "score": {
          "format": "bigint",
          "type": "integer"
        },
        "attempted_at": {
          "format": "timestamp with time zone",
          "type": "string"
        }
      },
      "type": "object"
    },
...
  },
  "parameters": {
...
    "preferReturn": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferCount": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "count=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferPost": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none",
        "resolution=ignore-duplicates",
        "resolution=merge-duplicates"
      ],
      "in": "header",
      "type": "string"
    },
    "select": {
      "name": "select",
      "description": "Filtering Columns",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "order": {
      "name": "order",
      "description": "Ordering",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "range": {
      "name": "Range",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "header",
      "type": "string"
    },
    "rangeUnit": {
      "name": "Range-Unit",
      "description": "Limiting and Pagination",
      "required": false,
      "default": "items",
      "in": "header",
      "type": "string"
    },
    "offset": {
      "name": "offset",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "limit": {
      "name": "limit",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "body.appointments": {
      "name": "appointments",
      "description": "appointments",
      "required": false,
      "in": "body",
      "schema": {
        "$ref": "#/definitions/appointments"
      }
    },
    "rowFilter.appointments.id": {
      "name": "id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.appointments.created_at": {
      "name": "created_at",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.appointments.doctor_id": {
      "name": "doctor_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.appointments.patient_uid": {
      "name": "patient_uid",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.appointments.patient_name": {
      "name": "patient_name",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.appointments.appointment_date": {
      "name": "appointment_date",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.appointments.start_time": {
      "name": "start_time",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.appointments.end_time": {
      "name": "end_time",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.appointments.status": {
      "name": "status",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "body.attempts": {
      "name": "attempts",
      "description": "attempts",
      "required": false,
      "in": "body",
      "schema": {
        "$ref": "#/definitions/attempts"
      }
    },
    "rowFilter.attempts.id": {
      "name": "id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.attempts.created_at": {
      "name": "created_at",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.attempts.patient_uid": {
      "name": "patient_uid",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.attempts.item_id": {
      "name": "item_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.attempts.alphabet_name": {
      "name": "alphabet_name",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.attempts.level_key": {
      "name": "level_key",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.attempts.score": {
      "name": "score",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.attempts.attempted_at": {
      "name": "attempted_at",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
  },
...
}